+++ src/net/server/channel/handlers/MTSHandler.java
@@ -21,19 +21,17 @@
  */
 package net.server.channel.handlers;
 
-import client.MapleCharacter;
-import client.MapleClient;
-import client.inventory.Equip;
-import client.inventory.Item;
-import client.inventory.MapleInventoryType;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.List;
+
 import net.AbstractMaplePacketHandler;
-import net.server.channel.Channel;
 import net.server.Server;
+import net.server.channel.Channel;
 import server.MTSItemInfo;
 import server.MapleInventoryManipulator;
 import server.MapleItemInformationProvider;
@@ -41,6 +39,11 @@ import tools.DatabaseConnection;
 import tools.MaplePacketCreator;
 import tools.Pair;
 import tools.data.input.SeekableLittleEndianAccessor;
+import client.MapleCharacter;
+import client.MapleClient;
+import client.inventory.Equip;
+import client.inventory.Item;
+import client.inventory.MapleInventoryType;
 
 public final class MTSHandler extends AbstractMaplePacketHandler {
 
@@ -68,15 +71,15 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
                 } else {
                     slea.readShort();
                 }
-                byte slot;
+                short slot;
                 short quantity;
                 if (itemtype != 1) {
                     if (itemid / 10000 == 207 || itemid / 10000 == 233) {
                         slea.skip(8);
                     }
-                    slot = (byte) slea.readInt();
+                    slot = (short) slea.readInt();
                 } else {
-                    slot = (byte) slea.readInt();
+                    slot = (short) slea.readInt();
                 }
                 if (itemtype != 1) {
                     if (itemid / 10000 == 207 || itemid / 10000 == 233) {
@@ -281,7 +284,7 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
                     if (rs.next()) {
                         Item i;
                         if (rs.getInt("type") != 1) {
-                            Item ii = new Item(rs.getInt("itemid"), (byte) 0, (short) rs.getInt("quantity"));
+                            Item ii = new Item(rs.getInt("itemid"), (short) 0, (short) rs.getInt("quantity"));
                             ii.setOwner(rs.getString("owner"));
                             ii.setPosition(c.getPlayer().getInventory(MapleItemInformationProvider.getInstance().getInventoryType(rs.getInt("itemid"))).getNextFreeSlot());
                             i = ii.copy();
@@ -568,7 +571,7 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
                     ResultSet rse = pse.executeQuery();
                     if (rse.next()) {
                         if (rse.getInt("type") != 1) {
-                            Item i = new Item(rse.getInt("itemid"), (byte) 0, (short) rse.getInt("quantity"));
+                            Item i = new Item(rse.getInt("itemid"), (short) 0, (short) rse.getInt("quantity"));
                             i.setOwner(rse.getString("owner"));
                             items.add(new MTSItemInfo((Item) i, rse.getInt("price"), rse.getInt("id"), rse.getInt("seller"), rse.getString("sellername"), rse.getString("sell_ends")));
                         } else {
@@ -628,7 +631,7 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
             rs = ps.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("type") != 1) {
-                    Item i = new Item(rs.getInt("itemid"), (byte) 0, (short) rs.getInt("quantity"));
+                    Item i = new Item(rs.getInt("itemid"), (short) 0, (short) rs.getInt("quantity"));
                     i.setOwner(rs.getString("owner"));
                     items.add(new MTSItemInfo((Item) i, rs.getInt("price"), rs.getInt("id"), rs.getInt("seller"), rs.getString("sellername"), rs.getString("sell_ends")));
                 } else {
@@ -686,7 +689,7 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
             rs = ps.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("type") != 1) {
-                    Item i = new Item(rs.getInt("itemid"), (byte) 0, (short) rs.getInt("quantity"));
+                    Item i = new Item(rs.getInt("itemid"), (short) 0, (short) rs.getInt("quantity"));
                     i.setOwner(rs.getString("owner"));
                     items.add(new MTSItemInfo((Item) i, rs.getInt("price"), rs.getInt("id"), rs.getInt("seller"), rs.getString("sellername"), rs.getString("sell_ends")));
                 } else {
@@ -777,7 +780,7 @@ public final class MTSHandler extends AbstractMaplePacketHandler {
             rs = ps.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("type") != 1) {
-                    Item i = new Item(rs.getInt("itemid"), (byte) 0, (short) rs.getInt("quantity"));
+                    Item i = new Item(rs.getInt("itemid"), (short) 0, (short) rs.getInt("quantity"));
                     i.setOwner(rs.getString("owner"));
                     items.add(new MTSItemInfo((Item) i, rs.getInt("price"), rs.getInt("id"), rs.getInt("seller"), rs.getString("sellername"), rs.getString("sell_ends")));
                 } else {
